requirements.txt
avro==1.11.1
avro_python3==1.10.2
confluent_kafka==2.1.1
python_binance==1.0.17
